+++ app/src/main/java/com/example/spender/ui/navigation/screens/balance_screens/add_spending_screen.kt
@@ -1,0 +1,28 @@
+package com.example.spender.ui.navigation.screens.balance_screens
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Button
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import com.example.spender.ui.navigation.nav_graphs.BalanceNavGraph
+import com.ramcosta.composedestinations.annotation.Destination
+import com.ramcosta.composedestinations.navigation.DestinationsNavigator
+
+@BalanceNavGraph
+@Destination
+@Composable
+fun AddSpendingScreen(
+    navigator: DestinationsNavigator
+) {
+    Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
+        Button(onClick = {
+            //navigator.
+        }) {
+            Text("Add screen")
+        }
+    }
+}
