+++ build.gradle
@@ -1,8 +1,8 @@
 buildscript {
 }// Top-level build file where you can add configuration options common to all sub-projects/modules.
     plugins {
-        id 'com.android.application' version '7.4.1' apply false
-        id 'com.android.library' version '7.4.1' apply false
+        id 'com.android.application' version '7.4.2' apply false
+        id 'com.android.library' version '7.4.2' apply false
         id 'com.google.android.libraries.mapsplatform.secrets-gradle-plugin' version '2.0.1' apply false
 
         // Navigation Destinations requires higher kotlin version for convenience
